requirements.txt
streamlit==1.45.1
streamlit-quill==0.0.3
pydicom
pillow
opencv-python-headless
pytesseract
python-docx
diff-match-patch
openai
